+++ producer/proto/reflect.go
@@ -9,23 +9,6 @@ import (
 	"github.com/netsampler/goflow2/v2/decoders/netflow"
 )
 
-type EndianType string
-type ProtoType string
-
-var (
-	BigEndian    EndianType = "big"
-	LittleEndian EndianType = "little"
-
-	ProtoString ProtoType = "string"
-	ProtoVarint ProtoType = "varint"
-
-	ProtoTypeMap = map[string]ProtoType{
-		string(ProtoString): ProtoString,
-		string(ProtoVarint): ProtoVarint,
-		"bytes":             ProtoString,
-	}
-)
-
 func GetBytes(d []byte, offset int, length int) []byte {
 	if length == 0 || offset < 0 {
 		return nil
@@ -153,22 +136,6 @@ func IsUInt(k reflect.Kind) bool {
 
 func IsInt(k reflect.Kind) bool {
 	return k == reflect.Int8 || k == reflect.Int16 || k == reflect.Int32 || k == reflect.Int64
-}
-
-// Structure to help the MapCustom functions
-// populate the protobuf data
-type MapConfigBase struct {
-	// Used if the field inside the protobuf exists
-	// also serves as the field when rendering with text
-	Destination string
-	Endianness  EndianType
-
-	// The following fields are used for mapping
-	// when the destination field does not exist
-	// inside the protobuf
-	ProtoIndex int32
-	ProtoType  ProtoType
-	ProtoArray bool
 }
 
 func MapCustomNetFlow(flowMessage *ProtoProducerMessage, df netflow.DataField, mapper *NetFlowMapper) error {
